requirements.txt
# requirements.txt
python-telegram-bot[job-queue]
python-dotenv
selenium
imageio[ffmpeg]